Create product first version

+++ src/main/java/com/ingemark/webshop/ProductRepository.java
@@ -3,5 +3,5 @@ package com.ingemark.webshop;
 import com.ingemark.webshop.model.ProductModel;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ProductRepository extends CrudRepository<ProductModel, Integer> {
+public interface ProductRepository extends CrudRepository<ProductModel, Long> {
 }
